+++ src/app/components/buckets-modal/buckets-modal.component.ts
@@ -1,7 +1,13 @@
 import { Component, Inject, OnDestroy, OnInit } from '@angular/core';
-import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
+import { MatDialog, MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { Board } from 'src/app/models/board';
+import User from 'src/app/models/user';
 import { BucketService } from 'src/app/services/bucket.service';
+import { PostService } from 'src/app/services/post.service';
+import { AddPostComponent } from 'src/app/components/add-post-modal/add-post.component';
+import Post from 'src/app/models/post';
+import { DialogInterface } from 'src/app/interfaces/dialog.interface';
+
 
 @Component({
   selector: 'app-buckets-modal',
@@ -11,6 +17,7 @@ import { BucketService } from 'src/app/services/bucket.service';
 export class BucketsModalComponent implements OnInit, OnDestroy {
 
   board: Board
+  user:User
 
   buckets: any
   activeBucket: any
@@ -22,8 +29,11 @@ export class BucketsModalComponent implements OnInit, OnDestroy {
   constructor(
     public dialogRef: MatDialogRef<BucketsModalComponent>,
     public bucketService: BucketService,
+    public postsService:PostService,
+    public dialog: MatDialog,
     @Inject(MAT_DIALOG_DATA) public data: any) {
     this.board = data.board
+    this.user = data.user
   }
 
   ngOnInit(): void {
@@ -62,5 +72,39 @@ export class BucketsModalComponent implements OnInit, OnDestroy {
     this.activeBucket = null
     this.buckets = []
     this.posts = []
+  }
+
+  addPost = (title: string, desc = '', left: number, top: number) => {
+    if (!this.activeBucket){
+      return;
+    }
+    const post: Post = {
+      postID: Date.now() + '-' + this.user.id,
+      title: title,
+      desc: desc,
+      tags: [],
+      userID: this.user.id,
+      boardID: this.board.boardID,
+      fabricObject: "{}",
+      timestamp: new Date().getTime(),
+      bucketOnly:true
+    }
+    this.postsService.create(post);
+    this.posts.push(post);
+    let ids = this.posts.map(post=>post.postID)
+    this.bucketService.update(this.activeBucket.bucketID,{posts:ids})
+
+  }
+
+  openAddPostDialog(){
+    const dialogData: DialogInterface = {
+      addPost: this.addPost,
+      top: 150,
+      left: 150,
+    }
+    this.dialog.open(AddPostComponent, {
+      width: '500px',
+      data: dialogData
+    });
   }
 }
