--- src/app/shared/bucket/bucket.component.html
@@ -1,31 +1,0 @@
-<mat-drawer-container class="drawer-container">
-    <mat-drawer #drawer class="drawer" mode="side">
-      <mat-toolbar>
-        <mat-form-field *ngIf="buckets && buckets.length > 0 && !loading" appearance="outline" style="font-size: 14px; width: 100%; margin-top: 20px">
-          <mat-select [(ngModel)]="activeBucket.name">
-            <mat-option value="{{bucket.name}}" *ngFor="let bucket of buckets" (click)="loadBucketPosts(bucket)">{{ bucket.name }}</mat-option>
-          </mat-select>
-        </mat-form-field>
-        <ng-template [ngIf]="buckets && buckets.length == 0">
-          <h3>No Buckets Available!</h3>
-        </ng-template>
-        <span style="flex: 1 1 auto"></span>
-        <button 
-          mat-icon-button 
-          matTooltip="Create Post in Bucket" 
-          (click)="openAddPostDialog()"
-          [disabled] = "buckets && buckets.length == 0">
-            <mat-icon>library_add</mat-icon>
-        </button>
-        <button mat-icon-button matTooltip="Close Bucket" (click)="close()">
-          <mat-icon>close</mat-icon>
-        </button>
-      </mat-toolbar>
-      <mat-drawer-content class="drawer-content">
-        <mat-spinner style="margin: 40% auto;" *ngIf="loading"></mat-spinner>
-        <div style='padding-top: 10px; height: calc(100% - 10px); overflow: scroll' *ngIf="!loading">
-          <app-bucket-post [post]="post" *ngFor="let post of posts"></app-bucket-post>
-        </div>
-      </mat-drawer-content>
-    </mat-drawer>
-</mat-drawer-container>
